+++ zcash_primitives/CHANGELOG.md
@@ -8,7 +8,7 @@ and this library adheres to Rust's notion of
 ## [Unreleased]
 ### Added
 - `zcash_primitives::sapling`:
-  - `keys::SharedSecret`
+  - `keys::{EphemeralSecretKey, EphemeralPublicKey, SharedSecret}`
   - `value`, containing types for handling Sapling note values and value
     commitments.